requirements.txt
# pip install -r requirements.txt

# Dependências básicas do sistema
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
requests>=2.28.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Dependências para Fase 4 - Interface de Revisão Humana
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Dependências opcionais para interface web
# streamlit>=1.28.0  # Para dashboard alternativo
# dash>=2.14.0       # Para dashboard interativo

# Dependências de desenvolvimento e teste
pytest>=7.4.0
pytest-asyncio>=0.21.0